stable: Finish dark theme toggle
