+++ backend/package.json
@@ -12,6 +12,6 @@
     "dependencies": {
       "dotenv": "^16.3.1",
       "express": "^4.18.2",
-      "mongoose": "^7.5.2"
+      "mongoose": "^7.5.3"
     }
   }
