+++ frontend/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 	<head>
 		<meta charset="UTF-8" />
-		<link rel="icon" type="image/svg+xml" href="" />
+		<!-- <link rel="icon" type="image/svg+xml" href="" /> -->
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>AI Personalities React App</title>
 		<script src="./src/components/DarkMode/DarkMode.ts"></script>
