+++ frontend/src/input.css
@@ -2,26 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-:root {
-	font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-	line-height: 1.5;
-	font-weight: 400;
-	max-width: 1280px;
-	margin: 0 auto;
-	padding: 2rem;
-	text-align: center;
-
-	color-scheme: light dark;
-	color: rgba(255, 255, 255, 0.87);
-	background-color: #242424;
-
-	font-synthesis: none;
-	text-rendering: optimizeLegibility;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-	-webkit-text-size-adjust: 100%;
-}
-
 a {
 	font-weight: 500;
 	color: #646cff;
@@ -33,11 +13,25 @@ a:hover {
 }
 
 body {
+	font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+	line-height: 1.5;
+	font-weight: 400;
+	text-align: center;
 	margin: 0;
+	padding: 2rem;
 	display: flex;
 	place-items: center;
-	min-width: 320px;
+	min-width: 100vw;
 	min-height: 100vh;
+	color-scheme: light dark;
+	color: rgba(255, 255, 255, 0.87);
+	background-color: #242424;
+
+	font-synthesis: none;
+	text-rendering: optimizeLegibility;
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+	-webkit-text-size-adjust: 100%;
 }
 
 .light-mode {
@@ -52,7 +46,7 @@ h1 {
 
 #theme-toggle {
 	border-radius: 8px;
-	border: 1px solid transparent;
+	border: 2px solid #888;
 	padding: 0.6em 1.2em;
 	font-size: 1em;
 	font-weight: 500;
@@ -69,21 +63,6 @@ h1 {
 #theme-toggle:focus,
 #theme-toggle:focus-visible {
 	outline: 4px auto -webkit-focus-ring-color;
-}
-
-@media (prefers-color-scheme: light) {
-	:root {
-		color: #213547;
-		background-color: #ffffff;
-	}
-
-	a:hover {
-		color: #747bff;
-	}
-
-	#theme-toggle {
-		background-color: #f9f9f9;
-	}
 }
 
 .logo {
