+++ frontend/src/main.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import '../src/output.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-	<React.StrictMode>
+	<StrictMode>
 		<App />
-	</React.StrictMode>
+	</StrictMode>
 );
